src/cryptor-trends-gemini-bot/pyproject.toml
[tool.poetry]
name = "cryptor-trends-gemini-bot"
version = "0.1.0"
description = ""
authors = ["Devops"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pika = "^1.3.2"
pymongo = "^4.6.3"
google-generativeai = "^0.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
